docker/ui/requirements.txt
streamlit
typing_extensions==4.9.0
psycopg2-binary
SQLAlchemy
python-dotenv